Add "Ticket" column (incompatible!); rework how stuff is sent to server
Reworking how controlls are being removed before tables are sent to the
server (so we do not loose data anymore when rename is active).

Incompatible with old tables, do not add activities to old topics!

No ticket marker, because this is for internal use only (the irony...).
Ticket: none
Release-Note: none